chore: move some helpers to core
