+++ packages/core/src/server/helper.ts
@@ -1,10 +1,11 @@
 import net from 'node:net';
+import { isIPv6 } from 'node:net';
+import os from 'node:os';
 import {
   DEFAULT_DEV_HOST,
   DEFAULT_PORT,
   color,
   deepmerge,
-  getUrlLabel,
   isFunction,
   logger,
   normalizeUrl,
@@ -287,4 +288,116 @@ export const getDevOptions = async ({
     https,
     liveReload,
   };
+};
+
+const getIpv4Interfaces = () => {
+  const interfaces = os.networkInterfaces();
+  const ipv4Interfaces: Map<string, os.NetworkInterfaceInfo> = new Map();
+
+  for (const key of Object.keys(interfaces)) {
+    for (const detail of interfaces[key]!) {
+      // 'IPv4' is in Node <= 17, from 18 it's a number 4 or 6
+      const familyV4Value = typeof detail.family === 'string' ? 'IPv4' : 4;
+
+      if (
+        detail.family === familyV4Value &&
+        !ipv4Interfaces.has(detail.address)
+      ) {
+        ipv4Interfaces.set(detail.address, detail);
+      }
+    }
+  }
+
+  return Array.from(ipv4Interfaces.values());
+};
+
+const isLoopbackHost = (host: string) => {
+  const loopbackHosts = [
+    'localhost',
+    '127.0.0.1',
+    '::1',
+    '0000:0000:0000:0000:0000:0000:0000:0001',
+  ];
+  return loopbackHosts.includes(host);
+};
+
+const getHostInUrl = (host: string) => {
+  if (isIPv6(host)) {
+    return host === '::' ? '[::1]' : `[${host}]`;
+  }
+  return host;
+};
+
+const concatUrl = ({
+  host,
+  port,
+  protocol,
+}: {
+  host: string;
+  port: number;
+  protocol: string;
+}) => `${protocol}://${host}:${port}`;
+
+const LOCAL_LABEL = 'Local:  ';
+const NETWORK_LABEL = 'Network:  ';
+
+export const getUrlLabel = (url: string) => {
+  try {
+    const { host } = new URL(url);
+    return isLoopbackHost(host) ? LOCAL_LABEL : NETWORK_LABEL;
+  } catch (err) {
+    return NETWORK_LABEL;
+  }
+};
+
+type AddressUrl = { label: string; url: string };
+
+export const getAddressUrls = ({
+  protocol = 'http',
+  port,
+  host,
+}: {
+  protocol?: string;
+  port: number;
+  host?: string;
+}): AddressUrl[] => {
+  if (host && host !== DEFAULT_DEV_HOST) {
+    return [
+      {
+        label: isLoopbackHost(host) ? LOCAL_LABEL : NETWORK_LABEL,
+        url: concatUrl({
+          port,
+          host: getHostInUrl(host),
+          protocol,
+        }),
+      },
+    ];
+  }
+
+  const ipv4Interfaces = getIpv4Interfaces();
+  const addressUrls: AddressUrl[] = [];
+  let hasLocalUrl = false;
+
+  for (const detail of ipv4Interfaces) {
+    if (isLoopbackHost(detail.address) || detail.internal) {
+      // avoid multiple prints of localhost
+      // https://github.com/web-infra-dev/rsbuild/discussions/1543
+      if (hasLocalUrl) {
+        continue;
+      }
+
+      addressUrls.push({
+        label: LOCAL_LABEL,
+        url: concatUrl({ host: 'localhost', port, protocol }),
+      });
+      hasLocalUrl = true;
+    } else {
+      addressUrls.push({
+        label: NETWORK_LABEL,
+        url: concatUrl({ host: detail.address, port, protocol }),
+      });
+    }
+  }
+
+  return addressUrls;
 };
