+++ packages/shared/src/minimize.ts
@@ -1,87 +1,7 @@
 import type { SwcJsMinimizerRspackPluginOptions } from '@rspack/core';
 import deepmerge from '../compiled/deepmerge/index.js';
-import type {
-  HTMLPluginOptions,
-  MinifyJSOptions,
-  NormalizedConfig,
-} from './types';
+import type { HTMLPluginOptions, NormalizedConfig } from './types';
 import { isObject } from './utils';
-
-function applyRemoveConsole(
-  options: MinifyJSOptions,
-  config: NormalizedConfig,
-) {
-  const { removeConsole } = config.performance;
-  const compressOptions =
-    typeof options.compress === 'boolean' ? {} : options.compress || {};
-
-  if (removeConsole === true) {
-    options.compress = {
-      ...compressOptions,
-      drop_console: true,
-    };
-  } else if (Array.isArray(removeConsole)) {
-    const pureFuncs = removeConsole.map((method) => `console.${method}`);
-    options.compress = {
-      ...compressOptions,
-      pure_funcs: pureFuncs,
-    };
-  }
-
-  return options;
-}
-
-function getTerserMinifyOptions(config: NormalizedConfig) {
-  const options: MinifyJSOptions = {
-    mangle: {
-      safari10: true,
-    },
-    format: {
-      ascii_only: config.output.charset === 'ascii',
-    },
-  };
-
-  if (config.output.legalComments === 'none') {
-    options.format!.comments = false;
-  }
-
-  const finalOptions = applyRemoveConsole(options, config);
-  return finalOptions;
-}
-
-export async function getHtmlMinifyOptions(
-  isProd: boolean,
-  config: NormalizedConfig,
-) {
-  if (
-    !isProd ||
-    !config.output.minify ||
-    !parseMinifyOptions(config).minifyHtml
-  ) {
-    return false;
-  }
-
-  const minifyJS: MinifyJSOptions = getTerserMinifyOptions(config);
-
-  const htmlMinifyDefaultOptions = {
-    removeComments: false,
-    useShortDoctype: true,
-    keepClosingSlash: true,
-    collapseWhitespace: true,
-    removeRedundantAttributes: true,
-    removeScriptTypeAttributes: true,
-    removeStyleLinkTypeAttributes: true,
-    removeEmptyAttributes: true,
-    minifyJS,
-    minifyCSS: true,
-    minifyURLs: true,
-  };
-
-  const htmlMinifyOptions = parseMinifyOptions(config).htmlOptions;
-  return typeof htmlMinifyOptions === 'object'
-    ? deepmerge(htmlMinifyDefaultOptions, htmlMinifyOptions)
-    : htmlMinifyDefaultOptions;
-}
 
 export const getSwcMinimizerOptions = (config: NormalizedConfig) => {
   const options: SwcJsMinimizerRspackPluginOptions = {};
