+++ packages/shared/tsconfig.json
@@ -2,6 +2,7 @@
   "extends": "@rsbuild/tsconfig/base",
   "compilerOptions": {
     "outDir": "./dist",
+    "declarationDir": "./dist-types",
     "baseUrl": "./",
     "rootDir": "src",
     "composite": true
